requirements.txt
numpy==1.21.4
pandas==1.3.4
Pillow==8.4.0
web.py==0.62
openpyxl==3.0.9
pipreqs==0.4.11
pyinstaller==4.7